backend/requirements.txt
Django>=3.1.2
djangorestframework>=3.12.1
drf_social_oauth2>=1.0.9
python-dotenv>=1.0.0
django-cors-headers>=3.7.0
